Create New Pudding Project

Creates a new R project and generates the folders needed for consistent Pudding analysis

create_project(name = "analysis", title = NULL, folder = getwd(),
  dirs = c("assets", "assets/data", "assets/data/open_data", "plots",
  "assets/data/processed_data", "assets/data/raw_data", "reports", "rmds"),
  packagedeps = "packrat", reset = FALSE, open = FALSE,
  defaultRmd = TRUE, brand = "pudding")

Arguments

name
File name for the project. Default: "analysis"

title
What the project does, one line, title case, Default: NULL

folder
Folder under which to create the new project, Default: getwd()

dirs
Character vector of new directories to create, Default: c("assets", "functions", "open_data", "plots", "processed_data", #' "raw_data", "reports", "rmds", "rscripts")

packagedeps
Set which tool you would like to use for package reproducibility, Default: 'packrat'

reset
Whether to reset the active project to your old project, Default: FALSE

open
Whether the new project should open a new session, Default: FALSE

defaultRmd
Whether to automatically create an Rmd file to get you started, Default: TRUE

brand
How should the default Rmd file be branded? Options: "pudding", "polygraph", Default: 'pudding'

Value

OUTPUT_DESCRIPTION

Details

DETAILS

See also

Examples

# NOT RUN {
folder <- tempdir()
createProject(
  name = "pudding-analysis", title = "My Next Great Story",
  folder = folder,
  packagedeps = "checkpoint",
  reset = TRUE,
  open = FALSE
)
list.files(file.path(folder, "pudding-analysis"))
unlink(file.path(folder, "pudding-analysis"))
# }